[win10] implemented uwp own StorageProvider.
