+++ cmake/treedata/windowsstore/subdirs.txt
@@ -6,7 +6,7 @@ xbmc/network/win10           network/win10
 xbmc/network/mdns            network/mdns
 xbmc/peripherals/bus/win10   peripherals/bus/win10
 xbmc/powermanagement/win10   powermanagement/win10
-xbmc/storage/windows         storage/windows
+xbmc/storage/win10           storage/win10
 xbmc/filesystem/win32        filesystem/win32
 xbmc/filesystem/win10        filesystem/win10
 xbmc/utils/win32             utils_win32
