+++ xbmc/powermanagement/windows/Win32PowerSyscall.cpp
@@ -19,6 +19,7 @@
  */
 
 #include "Win32PowerSyscall.h"
+#include "platform/win32/WIN32Util.h"
 #include "powermanagement/PowerManager.h"
 #include "utils/log.h"
 #include "utils/SystemInfo.h"
