+++ xbmc/storage/win10/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(SOURCES Win10StorageProvider.cpp)
+
+set(HEADERS Win10StorageProvider.h)
+
+core_add_library(storage_win10)
