+++ xbmc/storage/windows/Win32StorageProvider.h
@@ -18,12 +18,20 @@
  *  <http://www.gnu.org/licenses/>.
  *
  */
-
-#include <string>
+#pragma once
 #include <vector>
 
 #include "storage/IStorageProvider.h"
 #include "utils/Job.h"
+#include <Cfgmgr32.h>
+
+enum Drive_Types
+{
+  ALL_DRIVES = 0,
+  LOCAL_DRIVES,
+  REMOVABLE_DRIVES,
+  DVD_DRIVES
+};
 
 class CWin32StorageProvider : public IStorageProvider
 {
@@ -45,6 +53,10 @@ public:
 
   static void SetEvent() { xbevent = true; }
   static bool xbevent;
+
+private:
+  static void GetDrivesByType(VECSOURCES &localDrives, Drive_Types eDriveType=ALL_DRIVES, bool bonlywithmedia=false);
+  static DEVINST GetDrivesDevInstByDiskNumber(long DiskNumber);
 };
 
 class CDetectDisc : public CJob
